Fix string concatenation in geoip_client
It can happen that GeoIP records contain items like `"city": None`.
When we later go and try to concatenate the item's value to an already
existing string, this breaks for `None`, as `str` and `NoneType` are
different types and can't be concatenated without casting `None`.

This patch makes it so that we skip / suppress problematic entries under
the `"city"` and `"region_name"` keys in `geoip_location_str()`.
See also SeattleTestbed/seattlelib_v2#162, and note that seattlelib_v2
will need a similar patch applied.